[add]アカウント 認証

+++ database/migrations/2020_06_17_014658_create_screen_times_table.php
@@ -16,7 +16,7 @@ class CreateScreenTimesTable extends Migration
         Schema::create('screen_times', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->BigInteger('page_id');
-            $table->float('page_y',5,2);
+            $table->double('page_y', 10, 6);
             $table->tinyInteger('pc_tab_sp')->default(1);
             $table->timestamps();
         });
